studypulse/backend/requirements-railway.txt
# ── StudyPulse Backend - Railway Production Dependencies ──
# Minimal, Railway-optimized requirements (no heavy/problematic deps)
# CACHE BUST: 2026-02-17T10:30:00Z - Replaced passlib with direct bcrypt for compatibility

# ============ CORE FRAMEWORK ============
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
python-multipart>=0.0.6

# ============ DATABASE ============
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9  # Binary package - no pg_config needed

# ============ AUTHENTICATION & SECURITY ============
python-jose[cryptography]>=3.3.1
bcrypt==3.2.2  # Direct bcrypt usage (removed passlib for compatibility)
pydantic[email]>=2.7.0
pydantic-settings>=2.3.0
email-validator>=2.1.0
supabase==2.3.4,<2.4.0
gotrue>=2.0.0,<2.1.0
slowapi>=0.1.9

# ============ HTTP & NETWORKING ============
httpx>=0.24.0,<0.26  # Compatible with supabase 2.3.4

# ============ VECTOR SEARCH (Lightweight only) ============
hnswlib>=0.8.0
numpy>=1.24.0
fastembed>=0.4.0  # Required for PageIndex embeddings (lightweight ONNX)

# ============ CACHING ============
redis>=5.0.1

# ============ LOGGING & MONITORING ============
python-json-logger>=2.0.7
sentry-sdk[fastapi]>=1.40.0

# ============ CONFIGURATION ============
python-dotenv>=1.0.0

# ============ PDF PROCESSING (Basic only) ============
pypdf>=3.0.0
reportlab>=4.0.0

# ============ WEB SCRAPING (Basic only) ============
beautifulsoup4>=4.12.0
lxml>=5.0.0

# EXCLUDED for Railway (too heavy/problematic):
# - selenium (requires Chrome driver)
# - playwright (requires browser binaries)
# - qdrant-client (heavy ML dependencies, using PageIndex instead)
# - pdfplumber (optional, using pypdf only)
# - testing tools (pytest, locust, etc.)
